Pull up interpretation of HeapMap values into an abstract class
